Waypoint map cleanup
